+++ reservation/src/manager.rs
@@ -186,7 +186,7 @@ impl Rsvp for ReservationManager {
         let has_prev = !rsvps.is_empty() && rsvps[0].id == filter.cursor;
         // let start = if has_prev { 1 } else { 0 }; ==> usize::from(has_prev)
         let start = usize::from(has_prev);
-        let has_next = (rsvps.len() - start) as i32 > page_size;
+        let has_next = (rsvps.len() - start) as i64 > page_size;
         let end = if has_next {
             rsvps.len() - 1
         } else {
